Const.{h,cpp} を定数用、変数用のファイルに分割した
