+++ Ast/Declaration.h
@@ -2,7 +2,7 @@
 #define DECLARATION_H
 
 #include "Node.h"
-#include "Const.h"
+#include "ValueVariable.h"
 #include <llvm/IR/Module.h>
 #include <llvm/IR/IRBuilder.h>
 #include <iostream>
