+++ Ast/NodeAll.h
@@ -3,13 +3,14 @@
 
 #include "Node.h"
 #include "NodeList.h"
-#include "Const.h"
 #include "Control.h"
 #include "Declaration.h"
 #include "Definition.h"
 #include "Expression.h"
 #include "Jump.h"
 #include "Type.h"
+#include "ValueConst.h"
+#include "ValueVariable.h"
 
 #endif  // NODEALL_H
